Moving include folder according to http://www.ros.org/reps/rep-0122.html#include.
