+++ include/spinnaker_sdk_camera_driver/camera.h
@@ -1,8 +1,8 @@
 #ifndef CAMERA_HEADER
 #define CAMERA_HEADER
 
-#include "std_include.h"
-#include "serialization.h"
+#include "spinnaker_sdk_camera_driver/std_include.h"
+#include "spinnaker_sdk_camera_driver/serialization.h"
 #include <boost/archive/binary_oarchive.hpp>
 #include <boost/filesystem.hpp>
