+++ include/spinnaker_sdk_camera_driver/std_include.h
