+++ src/acquisition_node.cpp
@@ -1,4 +1,4 @@
-#include "capture.h"
+#include "spinnaker_sdk_camera_driver/capture.h"
 
 using namespace Spinnaker;
 using namespace Spinnaker::GenApi;
